refactor(ide): organize components by folder
- layout/ - activity-bar, ide-layout, title-bar, status-bar, mobile-menu
- sidebar/ - sidebar, source-control-view, file-icon
- editor/ - editor-pane, tabs, content, code-view, breadcrumbs, etc.
- terminal/ - terminal-panel, mock-terminal, commands, prompt
- command/ - command-palette
- shared/ - ide-types, split-editor-types, view-mode

No barrel files, direct imports only
