+++ apps/web/src/app/[locale]/(pages)/about/page.tsx
@@ -2,8 +2,8 @@ import { getTranslations, setRequestLocale } from "next-intl/server";
 import { AboutBio } from "@/components/about/about-bio";
 import { AboutHeader } from "@/components/about/about-header";
 import { ExperienceTimeline } from "@/components/about/experience-timeline";
-import { CodeView } from "@/components/ide/code-view";
-import { EditorContent } from "@/components/ide/editor-content";
+import { CodeView } from "@/components/ide/editor/code-view";
+import { EditorContent } from "@/components/ide/editor/editor-content";
 import { PageContent } from "@/components/shared/page-content";
 import { aboutCode } from "@/consts/code-content";
 import { createMetadata } from "@/lib/metadata";
