+++ apps/web/src/app/[locale]/(pages)/blog/page.tsx
@@ -3,8 +3,8 @@ import { Suspense } from "react";
 import { BlogHeader } from "@/components/blog/blog-header";
 import { BlogListLoader } from "@/components/blog/blog-list-loader";
 import { BlogListSkeleton } from "@/components/blog/blog-list-skeleton";
-import { CodeView } from "@/components/ide/code-view";
-import { EditorContent } from "@/components/ide/editor-content";
+import { CodeView } from "@/components/ide/editor/code-view";
+import { EditorContent } from "@/components/ide/editor/editor-content";
 import { PageContent } from "@/components/shared/page-content";
 import { blogCode } from "@/consts/code-content";
 import { createMetadata } from "@/lib/metadata";
