+++ apps/web/src/app/[locale]/page.tsx
@@ -7,8 +7,8 @@ import { HeroBio } from "@/components/home/hero-bio";
 import { HeroHeader } from "@/components/home/hero-header";
 import { HeroQuote } from "@/components/home/hero-quote";
 import { SocialNav } from "@/components/home/social-nav";
-import { CodeView } from "@/components/ide/code-view";
-import { EditorContent } from "@/components/ide/editor-content";
+import { CodeView } from "@/components/ide/editor/code-view";
+import { EditorContent } from "@/components/ide/editor/editor-content";
 import { PageContent } from "@/components/shared/page-content";
 import { welcomeCode } from "@/consts/code-content";
 import { createMetadata } from "@/lib/metadata";
