+++ apps/web/src/components/ide/command/command-palette.tsx
