+++ apps/web/src/components/ide/editor/code-editor.tsx
