+++ apps/web/src/components/ide/layout/ide-layout-skeleton.tsx
