+++ apps/web/src/components/ide/layout/ide-layout.tsx
@@ -9,21 +9,21 @@ import { navItems } from "@/consts/nav-items";
 import { useEditorGroups } from "@/hooks/use-editor-groups";
 import { useIdeKeyboardShortcuts } from "@/hooks/use-ide-keyboard-shortcuts";
 import { getBreadcrumbPath } from "@/lib/ide/breadcrumb";
-import { ActivityBar } from "@/components/ide/activity-bar";
-import { Breadcrumbs } from "@/components/ide/breadcrumbs";
-import { CommandPalette } from "@/components/ide/command-palette";
-import { EditorContentContextMenu } from "@/components/ide/editor-content-context-menu";
-import { EditorTabs } from "@/components/ide/editor-tabs";
-import { MobileMenu } from "@/components/ide/mobile-menu";
-import { Sidebar } from "@/components/ide/sidebar";
-import { SourceControlView } from "@/components/ide/source-control-view";
-import { SplitEditorView } from "@/components/ide/split-editor-view";
-import { StatusBar } from "@/components/ide/status-bar";
-import { TerminalPanel } from "@/components/ide/terminal-panel";
-import { TitleBar } from "@/components/ide/title-bar";
-import { type ViewMode, ViewModeProvider } from "@/components/ide/view-mode";
+import { ActivityBar } from "@/components/ide/layout/activity-bar";
+import { Breadcrumbs } from "@/components/ide/editor/breadcrumbs";
+import { CommandPalette } from "@/components/ide/command/command-palette";
+import { EditorContentContextMenu } from "@/components/ide/editor/editor-content-context-menu";
+import { EditorTabs } from "@/components/ide/editor/editor-tabs";
+import { MobileMenu } from "@/components/ide/layout/mobile-menu";
+import { Sidebar } from "@/components/ide/sidebar/sidebar";
+import { SourceControlView } from "@/components/ide/sidebar/source-control-view";
+import { SplitEditorView } from "@/components/ide/editor/split-editor-view";
+import { StatusBar } from "@/components/ide/layout/status-bar";
+import { TerminalPanel } from "@/components/ide/terminal/terminal-panel";
+import { TitleBar } from "@/components/ide/layout/title-bar";
+import { type ViewMode, ViewModeProvider } from "@/components/ide/shared/view-mode";
 import type { GitCommitItem } from "@/lib/github";
-import type { SidebarView } from "@/components/ide/ide-types";
+import type { SidebarView } from "@/components/ide/shared/ide-types";
 
 interface IdeLayoutProps {
   children: React.ReactNode;
