+++ apps/web/src/components/ide/layout/status-bar.tsx
