+++ apps/web/src/components/ide/layout/title-bar.tsx
