+++ apps/web/src/components/ide/shared/split-editor-types.ts
