+++ apps/web/src/components/ide/shared/view-mode.tsx
