+++ apps/web/src/components/ide/sidebar/file-icon.tsx
