+++ apps/web/src/components/ide/sidebar/sidebar.tsx
@@ -28,7 +28,7 @@ import {
   type FolderItem,
   type TreeItem,
 } from "@/consts/explorer-tree";
-import { FileIcon } from "@/components/ide/file-icon";
+import { FileIcon } from "./file-icon";
 
 interface SidebarProps {
   onOpenTab?: (href: string) => void;
