+++ apps/web/src/components/ide/terminal/terminal-panel.tsx
@@ -4,7 +4,7 @@ import { cn, Tooltip, TooltipContent, TooltipTrigger } from "@portfolio/ui";
 import { GripHorizontal, PanelBottomClose, Play, Terminal } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { MockTerminal } from "@/components/ide/mock-terminal";
+import { MockTerminal } from "./mock-terminal";
 
 const MIN_HEIGHT = 120;
 const MAX_HEIGHT = 600;
